fix: count the sudo marker on Unix only
Windows has no sudo. An elevated process with a stray SUDO_USER must
not divert its directories, refuse `pmg setup cert` or attribute the
audit to that name. The old euid check never fired on Windows, and
IsSudo now matches it.
