+++ internal/platform/privilege_test.go
@@ -16,28 +16,6 @@ func withPrivilege(t *testing.T, privileged bool) {
 	t.Cleanup(func() { IsPrivileged = orig })
 }
 
-func TestIsSudoCountsTheMarkerOnlyWhenPrivileged(t *testing.T) {
-	tests := []struct {
-		name       string
-		privileged bool
-		sudoUser   string
-		want       bool
-	}{
-		{name: "sudo from a person", privileged: true, sudoUser: "alice", want: true},
-		{name: "root without sudo", privileged: true, sudoUser: "", want: false},
-		{name: "a user who set the marker", privileged: false, sudoUser: "alice", want: false},
-		{name: "a user", privileged: false, sudoUser: "", want: false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			withPrivilege(t, tt.privileged)
-			t.Setenv("SUDO_USER", tt.sudoUser)
-			assert.Equal(t, tt.want, IsSudo())
-		})
-	}
-}
-
 func TestRequirePrivilege(t *testing.T) {
 	withPrivilege(t, true)
 	assert.NoError(t, RequirePrivilege("pmg setup install --system"))
